# Optimize checksum calculation on verify