Related to #614

- I modified the FileImpl::Verify checksum algo to calculate the checksum in pieces of **1024 bytes** to save on cpu cycles and IO requests.
- Around **2.6x** fast (tested on 16 core AMD EPYC 7B13 Gitpod)

|**Size** | **Previous** | **Optimized**|
|-----------|------------------|-------------------|
31 MB    |       0.73s     |     **0.27s**   |
4 GB      |      96.84s    |     **38.17s**   |